refactor: cycle api 관련 서비스 분리
