+++ backend/smody/src/main/java/com/woowacourse/smody/cycle/controller/CycleController.java
@@ -10,8 +10,7 @@ import com.woowacourse.smody.cycle.dto.InProgressCycleResponse;
 import com.woowacourse.smody.cycle.dto.ProgressRequest;
 import com.woowacourse.smody.cycle.dto.ProgressResponse;
 import com.woowacourse.smody.cycle.dto.StatResponse;
-import com.woowacourse.smody.cycle.service.CycleQueryService;
-import com.woowacourse.smody.cycle.service.CycleService;
+import com.woowacourse.smody.cycle.service.CycleApiService;
 import com.woowacourse.smody.db_support.PagingParams;
 import java.net.URI;
 import java.time.LocalDateTime;
@@ -31,14 +30,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class CycleController {
 
-    private final CycleService cycleService;
-    private final CycleQueryService cycleQueryService;
+    private final CycleApiService cycleApiService;
 
     @PostMapping
     @RequiredLogin
     public ResponseEntity<Void> create(@LoginMember TokenPayload tokenPayload,
                                        @RequestBody CycleRequest cycleRequest) {
-        Long cycleId = cycleService.create(tokenPayload, cycleRequest);
+        Long cycleId = cycleApiService.create(tokenPayload, cycleRequest);
         return ResponseEntity.created(URI.create("/cycles/" + cycleId)).build();
     }
 
@@ -47,27 +45,27 @@ public class CycleController {
     public ResponseEntity<ProgressResponse> increase(@LoginMember TokenPayload tokenPayload,
                                                      @ModelAttribute ProgressRequest progressRequest) {
         progressRequest.setProgressTime(LocalDateTime.now());
-        ProgressResponse progressResponse = cycleService.increaseProgress(tokenPayload, progressRequest);
+        ProgressResponse progressResponse = cycleApiService.increaseProgress(tokenPayload, progressRequest);
         return ResponseEntity.ok(progressResponse);
     }
 
     @GetMapping(value = "/me")
     @RequiredLogin
-    public ResponseEntity<List<InProgressCycleResponse>> findAllInProgressOfMine(@LoginMember TokenPayload tokenPayload,
-                                                                                 @ModelAttribute PagingParams pagingParams) {
+    public ResponseEntity<List<InProgressCycleResponse>> findAllInProgressByMe(@LoginMember TokenPayload tokenPayload,
+                                                                               @ModelAttribute PagingParams pagingParams) {
         return ResponseEntity.ok(
-                cycleQueryService.findInProgressOfMine(tokenPayload, LocalDateTime.now(), pagingParams));
+                cycleApiService.findInProgressByMe(tokenPayload, LocalDateTime.now(), pagingParams));
     }
 
     @GetMapping("/{cycleId}")
-    public ResponseEntity<CycleResponse> findById(@PathVariable Long cycleId) {
-        return ResponseEntity.ok(cycleQueryService.findById(cycleId));
+    public ResponseEntity<CycleResponse> findWithSuccessCountById(@PathVariable Long cycleId) {
+        return ResponseEntity.ok(cycleApiService.findWithSuccessCountById(cycleId));
     }
 
     @GetMapping("/me/stat")
     @RequiredLogin
     public ResponseEntity<StatResponse> searchStat(@LoginMember TokenPayload tokenPayload) {
-        return ResponseEntity.ok(cycleQueryService.searchStat(tokenPayload));
+        return ResponseEntity.ok(cycleApiService.searchStat(tokenPayload));
     }
 
     @GetMapping("/me/{challengeId}")
@@ -77,7 +75,7 @@ public class CycleController {
             @PathVariable Long challengeId,
             @ModelAttribute PagingParams pagingParams) {
         return ResponseEntity.ok(
-                cycleQueryService.findAllByMemberAndChallenge(tokenPayload, challengeId, pagingParams));
+                cycleApiService.findAllByMemberAndChallenge(tokenPayload, challengeId, pagingParams));
     }
 }
