+++ backend/smody/src/main/java/com/woowacourse/smody/cycle/service/CycleService.java
@@ -2,29 +2,21 @@ package com.woowacourse.smody.cycle.service;
 
 import static java.util.stream.Collectors.toList;
 
-import com.woowacourse.smody.auth.dto.TokenPayload;
 import com.woowacourse.smody.challenge.domain.Challenge;
+import com.woowacourse.smody.challenge.domain.ChallengingRecord;
 import com.woowacourse.smody.challenge.service.ChallengeService;
 import com.woowacourse.smody.cycle.domain.Cycle;
-import com.woowacourse.smody.cycle.domain.CycleCreateEvent;
-import com.woowacourse.smody.cycle.domain.CycleProgressEvent;
 import com.woowacourse.smody.cycle.domain.Progress;
-import com.woowacourse.smody.cycle.dto.CycleRequest;
-import com.woowacourse.smody.cycle.dto.ProgressRequest;
-import com.woowacourse.smody.cycle.dto.ProgressResponse;
 import com.woowacourse.smody.cycle.repository.CycleRepository;
 import com.woowacourse.smody.db_support.PagingParams;
 import com.woowacourse.smody.exception.BusinessException;
 import com.woowacourse.smody.exception.ExceptionData;
-import com.woowacourse.smody.image.domain.Image;
-import com.woowacourse.smody.image.strategy.ImageStrategy;
 import com.woowacourse.smody.member.domain.Member;
 import com.woowacourse.smody.member.service.MemberService;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -36,23 +28,17 @@ public class CycleService {
     private final CycleRepository cycleRepository;
     private final MemberService memberService;
     private final ChallengeService challengeService;
-    private final ImageStrategy imageStrategy;
-    private final ApplicationEventPublisher applicationEventPublisher;
 
     @Transactional
-    public Long create(TokenPayload tokenPayload, CycleRequest cycleRequest) {
-        Member member = memberService.search(tokenPayload.getId());
-        Challenge challenge = challengeService.search(cycleRequest.getChallengeId());
+    public Cycle create(Long memberId, Long challengeId, LocalDateTime startTime) {
+        Member member = memberService.search(memberId);
+        Challenge challenge = challengeService.search(challengeId);
         Optional<Cycle> optionalCycle = cycleRepository.findRecent(member, challenge);
 
-        LocalDateTime startTime = cycleRequest.getStartTime();
         if (optionalCycle.isPresent()) {
             startTime = calculateNewStartTime(startTime, optionalCycle.get());
         }
-        Cycle save = cycleRepository.save(new Cycle(member, challenge, Progress.NOTHING, startTime));
-
-        applicationEventPublisher.publishEvent(new CycleCreateEvent(save));
-        return save.getId();
+        return cycleRepository.save(new Cycle(member, challenge, Progress.NOTHING, startTime));
     }
 
     private LocalDateTime calculateNewStartTime(LocalDateTime startTime, Cycle cycle) {
@@ -63,23 +49,6 @@ public class CycleService {
             return cycle.getStartTime().plusDays(Cycle.DAYS);
         }
         return startTime;
-    }
-
-    @Transactional
-    public ProgressResponse increaseProgress(TokenPayload tokenPayload, ProgressRequest progressRequest) {
-        Cycle cycle = searchWithLock(progressRequest.getCycleId());
-        validateAuthorizedMember(tokenPayload, cycle);
-        Image progressImage = new Image(progressRequest.getProgressImage(), imageStrategy);
-        cycle.increaseProgress(progressRequest.getProgressTime(), progressImage, progressRequest.getDescription());
-
-        applicationEventPublisher.publishEvent(new CycleProgressEvent(cycle));
-        return new ProgressResponse(cycle.getProgress());
-    }
-
-    private void validateAuthorizedMember(TokenPayload tokenPayload, Cycle cycle) {
-        if (!cycle.matchMember(tokenPayload.getId())) {
-            throw new BusinessException(ExceptionData.UNAUTHORIZED_MEMBER);
-        }
     }
 
     public int countSuccess(Cycle cycle) {
@@ -125,5 +94,16 @@ public class CycleService {
                 .stream()
                 .filter(cycle -> cycle.isInProgress(searchTime))
                 .collect(toList());
+    }
+
+    public List<Cycle> findAllByMemberAndChallengeAndFilter(Long memberId, Long challengeId, PagingParams pagingParams) {
+        return cycleRepository.findAllByMemberAndChallengeAndFilter(memberId, challengeId, pagingParams);
+    }
+
+    public List<ChallengingRecord> findAllChallengingRecordByMember(Member member) {
+        return cycleRepository.findAllChallengingRecordByMemberAfterTime(
+                member.getId(),
+                LocalDateTime.now().minusDays(Cycle.DAYS)
+        );
     }
 }
