+++ backend/smody/src/test/java/com/woowacourse/smody/cycle/controller/CycleControllerTest.java
@@ -52,7 +52,7 @@ public class CycleControllerTest extends ControllerTest {
         // given
         Long cycleId = 1L;
         CycleRequest request = new CycleRequest(LocalDateTime.now(), 1L);
-        given(cycleService.create(any(TokenPayload.class), any(CycleRequest.class))).willReturn(cycleId);
+        given(cycleApiService.create(any(TokenPayload.class), any(CycleRequest.class))).willReturn(cycleId);
         String token = jwtTokenProvider.createToken(new TokenPayload(1L));
 
         // when
@@ -95,7 +95,7 @@ public class CycleControllerTest extends ControllerTest {
         // given
         String token = jwtTokenProvider.createToken(new TokenPayload(1L));
         ProgressResponse response = new ProgressResponse(2);
-        given(cycleService.increaseProgress(any(TokenPayload.class), any(ProgressRequest.class)))
+        given(cycleApiService.increaseProgress(any(TokenPayload.class), any(ProgressRequest.class)))
                 .willReturn(response);
 
         // when
@@ -125,7 +125,7 @@ public class CycleControllerTest extends ControllerTest {
     void increaseProgress_400() throws Exception {
         // given
         String token = jwtTokenProvider.createToken(new TokenPayload(1L));
-        given(cycleService.increaseProgress(any(TokenPayload.class), any(ProgressRequest.class)))
+        given(cycleApiService.increaseProgress(any(TokenPayload.class), any(ProgressRequest.class)))
                 .willThrow(new BusinessException(ExceptionData.INVALID_PROGRESS_TIME));
 
         // when
@@ -141,7 +141,7 @@ public class CycleControllerTest extends ControllerTest {
     void increaseProgress_403() throws Exception {
         // given
         String token = jwtTokenProvider.createToken(new TokenPayload(1L));
-        given(cycleService.increaseProgress(any(TokenPayload.class), any(ProgressRequest.class)))
+        given(cycleApiService.increaseProgress(any(TokenPayload.class), any(ProgressRequest.class)))
                 .willThrow(new BusinessException(ExceptionData.UNAUTHORIZED_MEMBER));
 
         // when
@@ -161,7 +161,7 @@ public class CycleControllerTest extends ControllerTest {
         List<InProgressCycleResponse> inProgressCycleResponses = List.of(
                 new InProgressCycleResponse(1L, 1L, "미라클 모닝", 2, now, 3, 0, 1),
                 new InProgressCycleResponse(2L, 2L, "오늘의 운동", 1, now, 3, 0, 1));
-        given(cycleQueryService.findInProgressOfMine(
+        given(cycleApiService.findInProgressByMe(
                 any(TokenPayload.class), any(LocalDateTime.class), any(PagingParams.class)
         )).willReturn(inProgressCycleResponses);
 
@@ -195,7 +195,7 @@ public class CycleControllerTest extends ControllerTest {
                 cycleId, 1L, "미라클 모닝", 2, LocalDateTime.now(), 3, "미라클 모닝입니다", 0, 1,
                 List.of(new CycleDetailResponse(LocalDateTime.now(), "image1", "인증 내용1"),
                         new CycleDetailResponse(LocalDateTime.now(), "image2", "인증 내용2")));
-        given(cycleQueryService.findById(cycleId))
+        given(cycleApiService.findWithSuccessCountById(cycleId))
                 .willReturn(cycleResponse);
 
         // when
@@ -230,7 +230,7 @@ public class CycleControllerTest extends ControllerTest {
     void findById_404() throws Exception {
         // given
         long cycleId = 1L;
-        given(cycleQueryService.findById(cycleId))
+        given(cycleApiService.findWithSuccessCountById(cycleId))
                 .willThrow(new BusinessException(ExceptionData.NOT_FOUND_CYCLE));
 
         // when
@@ -246,7 +246,7 @@ public class CycleControllerTest extends ControllerTest {
         // given
         String token = jwtTokenProvider.createToken(new TokenPayload(1L));
         StatResponse statResponse = new StatResponse(35, 5);
-        given(cycleQueryService.searchStat(any(TokenPayload.class)))
+        given(cycleApiService.searchStat(any(TokenPayload.class)))
                 .willReturn(statResponse);
 
         // when
@@ -296,7 +296,7 @@ public class CycleControllerTest extends ControllerTest {
                         )
                 ))
         );
-        given(cycleQueryService.findAllByMemberAndChallenge(any(TokenPayload.class), eq(1L), any(PagingParams.class)))
+        given(cycleApiService.findAllByMemberAndChallenge(any(TokenPayload.class), eq(1L), any(PagingParams.class)))
                 .willReturn(filteredCycleHistoryRespons);
 
         // when
@@ -355,7 +355,7 @@ public class CycleControllerTest extends ControllerTest {
                         )
                 ))
         );
-        given(cycleQueryService.findAllByMemberAndChallenge(any(TokenPayload.class), eq(1L), any(PagingParams.class)))
+        given(cycleApiService.findAllByMemberAndChallenge(any(TokenPayload.class), eq(1L), any(PagingParams.class)))
                 .willReturn(filteredCycleHistoryRespons);
 
         // when
